requirements.txt
<<<<<<< HEAD
black==24.2.0
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.1
=======
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.29.5
azure-identity==1.15.0
black==23.11.0
blinker==1.7.0
cachelib==0.12.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
Deprecated==1.2.14
et-xmlfile==1.1.0
exceptiongroup==1.1.3
flake8==6.1.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Session==0.4.1
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
identity==0.2.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
mccabe==0.7.0
microsoft-kiota-abstractions==1.0.0
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-http==1.0.0
microsoft-kiota-serialization-json==1.0.0
microsoft-kiota-serialization-text==1.0.0
msal==1.24.1
msal-extensions==1.0.0
msgraph-core==1.0.0a4
msgraph-sdk==1.0.0
multidict==6.0.4
mypy-extensions==1.0.0
openpyxl==3.1.2
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
packaging==23.2
pathspec==0.11.2
pendulum==2.1.2
platformdirs==4.1.0
pluggy==1.3.0
portalocker==2.8.2
psycopg==3.1.13
psycopg-binary==3.1.13
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytzdata==2020.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
std-uritemplate==0.0.46
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
>>>>>>> 72bf0d83f9556e84e035131127f4d23bb38abc65
